requirements.txt
langchain
langchain-community
langchain-huggingface
sentence-transformers
chromadb
python-dotenv
requests
openai
tiktoken
streamlit